feat(T13): Add localy done and undo task functions. Organize code catalogs structure/
