+++ src/app/components/shopping-list/shopping-list.component.spec.ts
