+++ src/app/components/todo-task/todo-task.component.ts
@@ -4,8 +4,8 @@ import {
   ViewChild,
   ViewEncapsulation,
 } from '@angular/core';
-import { TaskService } from 'src/app/services/task.service';
-import { Task } from 'src/app/models/task';
+import { TaskService } from 'src/app/shared/services/task.service';
+import { Task } from 'src/app/shared/interfaces/task';
 
 @Component({
   selector: 'app-todo-task',
