+++ src/app/config/firebase.config.ts
@@ -1,4 +1,4 @@
-import { Config } from '../models/config';
+import { Config } from '../shared/interfaces/config';
 
 export const config: Config = {
   firebase: {
