+++ src/app/shared/interfaces/config.ts
