+++ src/app/shared/pipes/transform-task.pipe.ts
